Feed a canonical int32 param to the raw-inputs simulate test
`test_simulate_using_raw_inputs` calls `simulate` directly, bypassing the
public `Model.simulate` canonicalization, so its `flat_params` must already
carry canonical dtypes. `jnp.asarray(0)` is `int64` under x64, but the
params-leaf contract is `int32`; `_invert_joint`'s beartype check rejected it
whenever its O(1) sampling drew that key. Pin the leaf to `int32` — the dtype
`cast_params_to_canonical_dtypes` produces on the public path.

+++ tests/simulation/test_simulate.py
@@ -77,7 +77,9 @@ def test_simulate_using_raw_inputs(simulate_inputs):
                     "H__discount_factor": jnp.asarray(1.0),
                     "utility__disutility_of_work": jnp.asarray(1.0),
                     "working_life__next_wealth__interest_rate": jnp.asarray(0.05),
-                    "next_regime__final_age_alive": jnp.asarray(0),
+                    # `simulate` consumes already-canonical params; an `int` leaf
+                    # is `int32` (a bare `jnp.asarray(0)` would be `int64` under x64).
+                    "next_regime__final_age_alive": jnp.asarray(0, dtype=jnp.int32),
                 }
             ),
             "dead": MappingProxyType({}),
